Initial node js typescript build
