+++ .gitignore
@@ -1,4 +1,4 @@
-server/private_credentials
+private_credentials/
 
 # compiled output
 /tmp
