+++ package.json
@@ -7,7 +7,8 @@
     "build": "webpack --env=production --progress --p",
     "postinstall": "typings install && npm run build",
     "prodBuild": "npm run build && node server/app.js",
-    "start": "concurrently \"webpack --env=development --watch --colors\" \"node server/app.js\""
+    "runTypescriptBuild": "tsc && node dist/app",
+    "start": "concurrently \"webpack --env=development --watch --colors\" \"npm run runTypescriptBuild && node dist/app\""
   },
   "author": "Juan Cortez",
   "repository": {
